+++ src/models/convnext/convnextv2.py
@@ -11,8 +11,6 @@
 from __future__ import annotations
 
 '''
-Ok, time to figure out what this architecture will look like, oh gosh.
-
 V1 Paper -> https://arxiv.org/pdf/2201.03545
 V2 Paper -> https://arxiv.org/pdf/2301.00808
 '''
@@ -31,7 +29,7 @@ class Block(nn.Module):
         dim (int): Number of input channels.
         drop_path (float): Stochastic depth rate. Default: 0.0
     """
-    def __init__(self, dim, drop_path=0.):
+    def __init__(self, dim: int, drop_path: float=0.):
         super().__init__()
         self.dwconv = nn.Conv2d(dim, dim, kernel_size=7, padding=3, groups=dim) # depthwise conv
         self.norm = LayerNorm(dim, eps=1e-6)
@@ -41,9 +39,9 @@ class Block(nn.Module):
         self.pwconv2 = nn.Linear(4 * dim, dim)
         self.drop_path = DropPath(drop_path) if drop_path > 0. else nn.Identity()
 
-    def forward(self, x):
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
         input = x
-        x = self.dwconv(x)
+        x = self.dwconv(x) # pseudo-sparse mask would go before and after this?
         x = x.permute(0, 2, 3, 1) # (N, C, H, W) -> (N, H, W, C)
         x = self.norm(x)
         x = self.pwconv1(x)
@@ -68,31 +66,42 @@ class ConvNeXtV2(nn.Module):
     """
     def __init__(
         self,
-        in_chans=3, 
-        num_classes=1000, 
-        depths=[3, 3, 9, 3],
-        dims=[96, 192, 384, 768], 
-        drop_path_rate=0.,
-        head_init_scale=1.
+        in_chans: int=3, 
+        num_classes: int=300, # default changed from 1000 to 300 for this project
+        depths: list[int]=[3, 3, 9, 3],
+        dims: list[int]=[96, 192, 384, 768], 
+        drop_path_rate: float=0.,
+        head_init_scale: float=1.
     ):
+        # initialization
         super().__init__()
         self.depths = depths
         self.downsample_layers = nn.ModuleList() # stem and 3 intermediate downsampling conv layers
+
+        # stem ("patchify" layer)
         stem = nn.Sequential(
             nn.Conv2d(in_chans, dims[0], kernel_size=4, stride=4),
-            LayerNorm(dims[0], eps=1e-6, data_format="channels_first")
+            LayerNorm(dims[0], eps=1e-6, channels_last=False)
         )
+
         self.downsample_layers.append(stem)
+
+        # downsample layers before passing into each block
         for i in range(3):
             downsample_layer = nn.Sequential(
-                    LayerNorm(dims[i], eps=1e-6, data_format="channels_first"),
+                    LayerNorm(dims[i], eps=1e-6, channels_last=False),
                     nn.Conv2d(dims[i], dims[i+1], kernel_size=2, stride=2),
             )
             self.downsample_layers.append(downsample_layer)
 
         self.stages = nn.ModuleList() # 4 feature resolution stages, each consisting of multiple residual blocks
+        
+        # drop path rates
         dp_rates=[x.item() for x in torch.linspace(0, drop_path_rate, sum(depths))] 
+        
         cur = 0
+        
+        # ConvNeXt V2 blocks
         for i in range(4):
             stage = nn.Sequential(
                 *[Block(dim=dims[i], drop_path=dp_rates[cur + j]) for j in range(depths[i])]
@@ -100,31 +109,33 @@ class ConvNeXtV2(nn.Module):
             self.stages.append(stage)
             cur += depths[i]
 
-        self.norm = nn.LayerNorm(dims[-1], eps=1e-6) # final norm layer
+        # final norm layer
+        self.norm = nn.LayerNorm(dims[-1], eps=1e-6)
         self.head = nn.Linear(dims[-1], num_classes)
 
+        # weight initialization 
         self.apply(self._init_weights)
         self.head.weight.data.mul_(head_init_scale)
         self.head.bias.data.mul_(head_init_scale)
 
         # for semantic segmentation skip connection
         self.seg_norms = nn.ModuleList(
-            [LayerNorm(dims[i], eps=1e-6, data_format="channels_first") for i in range (4)]
+            [LayerNorm(dims[i], eps=1e-6, channels_last=False) for i in range (4)]
         )
 
-    def _init_weights(self, m):
+    def _init_weights(self, m: nn.Module) -> None:
         if isinstance(m, (nn.Conv2d, nn.Linear)):
             trunc_normal_(m.weight, std=.02)
             nn.init.constant_(m.bias, 0) # type: ignore
 
-    def forward_features(self, x):
+    def forward_features(self, x: torch.Tensor) -> torch.Tensor:
         for i in range(4):
             x = self.downsample_layers[i](x)
             x = self.stages[i](x)
         return self.norm(x.mean([-2, -1])) # global average pooling, (N, C, H, W) -> (N, C)
 
     # create an updated forward_features for segmentation task
-    def forward_features_seg(self, x):
+    def forward_features_seg(self, x: torch.Tensor) -> tuple:
         outs = []
         for i in range(4):
             x = self.downsample_layers[i](x)
@@ -132,7 +143,8 @@ class ConvNeXtV2(nn.Module):
             outs.append(self.seg_norms[i](x))
         return tuple(outs)
 
-    def forward(self, x):
+    # base ConvNeXt V2 is configured for image classification
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
         x = self.forward_features(x)
         x = self.head(x)
         return x